Host의 winnigs, totalRank -> Person으로 이동 & View 패키지 생성
