+++ Lotto/src/main/kotlin/Host.kt
@@ -1,13 +1,5 @@
 class Host() {
     val winLotto = setWinLotto()
-    var winnigs = 0
-    val totalRank = mutableMapOf(
-        Rank.FIRST to 0,
-        Rank.SECOND to 0,
-        Rank.THIRD to 0,
-        Rank.FOURTH to 0,
-        Rank.LOSE to 0,
-    )
 
     fun setWinLotto(): Lotto {
         return Lotto(Lotto.selectLottoNumber())
@@ -17,14 +9,14 @@ class Host() {
         return winLotto.getLottoNumber().intersect(lotto.getLottoNumber()).count()
     }
 
-    fun countRank(count: Int) {
+    fun countRank(count: Int, person: Person) {
         val rank = Rank.getRank(count)
-        totalRank[rank] = totalRank.getValue(rank) + 1
+        person.totalRank[rank] = person.totalRank.getValue(rank) + 1
     }
 
-    fun calculateWinnigs() {
-        totalRank.forEach { (rank, count) ->
-            winnigs += rank.money * count
+    fun calculateWinnigs(person: Person) {
+        person.totalRank.forEach { (rank, count) ->
+            person.winnigs += rank.money * count
         }
     }
 }
