+++ Lotto/src/main/kotlin/LottoController.kt
@@ -1,3 +1,5 @@
+import View.*
+
 class LottoController {
     fun process() {
         val money = inputMoney() // View에게 금액을 입력받도록 명령
@@ -5,9 +7,9 @@ class LottoController {
         val person = Person(money)
         buyLottos(person) // Person이 Lottos를 구매하도록 명령
         printLotto(person.lottos) // Person이 Lottos를 출력하도록 명령
-        processLotto(host, person.lottos) // Host가 Lottos를 체크하도록 명령
-        host.calculateWinnigs() // Host가 총 당첨금을 계산하도록 명령
-        result(host) // 결과를 출력
+        processLotto(host, person) // Host가 Lottos를 체크하도록 명령
+        host.calculateWinnigs(person) // Host가 총 당첨금을 계산하도록 명령
+        result(person) // 결과를 출력
     }
     fun buyLottos(person: Person) {
         val manualBuy = selectBuyOption()
@@ -21,10 +23,10 @@ class LottoController {
             person.lottoBuy(lottoNumber)
         }
     }
-    fun processLotto(host: Host, lottos: Lottos) {
-        lottos.getList().forEach { lotto: Lotto -> host.countRank(host.checkLotto(lotto)) }
+    fun processLotto(host: Host, person: Person) {
+        person.lottos.getList().forEach { lotto: Lotto -> host.countRank(host.checkLotto(lotto), person) }
     }
-    fun result(host: Host) {
-        printResult(host)
+    fun result(person: Person) {
+        printResult(person)
     }
 }
