+++ Lotto/src/main/kotlin/View/ResultView.kt
@@ -1,4 +1,7 @@
+package View
 
+import Lottos
+import Person
 
 fun printLotto(lottos: Lottos) {
     println("사용자가 구매 한 로또 번호는 다음과 같습니다.")
@@ -7,12 +10,12 @@ fun printLotto(lottos: Lottos) {
     }
 }
 
-fun printResult(host: Host) {
+fun printResult(person: Person) {
     println("--------------------")
     println("당첨 결과")
-    host.totalRank.forEach { (rank, count) ->
+    person.totalRank.forEach { (rank, count) ->
         println("$rank : ${rank.money}KW -> ${count}개")
     }
     println("--------------------")
-    println("총 당첨금 : ${host.winnigs}KW")
+    println("총 당첨금 : ${person.winnigs}KW")
 }
